fzj 23th commit at 2021/10/05 fix post id type

+++ lianxiao-simple-server/src/main/java/com/lianxiao/demo/simpleserver/dao/PostRepository.java
@@ -3,5 +3,5 @@ package com.lianxiao.demo.simpleserver.dao;
 import com.lianxiao.demo.simpleserver.model.Post;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
-public interface PostRepository extends ElasticsearchRepository<Post, Integer> {
+public interface PostRepository extends ElasticsearchRepository<Post, Long> {
 }
